add TableMetaDataLoader.loadAll()
