+++ sharding-proxy/sharding-proxy-backend/src/main/java/org/apache/shardingsphere/shardingproxy/backend/schema/impl/TransparentSchema.java
@@ -51,7 +51,7 @@ public final class TransparentSchema extends LogicSchema {
     
     private ShardingSphereMetaData createMetaData() throws SQLException {
         DataSourceMetas dataSourceMetas = new DataSourceMetas(LogicSchemas.getInstance().getDatabaseType(), getDatabaseAccessConfigurationMap());
-        TableMetas tableMetas = createTableMetaDataInitializerEntry(dataSourceMetas).load(shardingRule);
+        TableMetas tableMetas = createTableMetaDataInitializerEntry(dataSourceMetas).loadAll(shardingRule);
         return new ShardingSphereMetaData(dataSourceMetas, tableMetas);
     }
 }
